microservicio1/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
